+++ src/rnnoise/celt_lpc.h
@@ -37,9 +37,8 @@
 
 #define LPC_ORDER 24
 
-void rnn_lpc(opus_val16 *_lpc, const opus_val32 *ac, int p);
+void rnn_lpc(opus_val16* _lpc, const opus_val32* ac, int p);
 
-int rnn_autocorr(const opus_val16 *x, opus_val32 *ac,
-         const opus_val16 *window, int overlap, int lag, int n);
+int rnn_autocorr(const opus_val16* x, opus_val32* ac, const opus_val16* window, int overlap, int lag, int n);
 
 #endif /* PLC_H */